Resolvendo Bugs de inicialização da aplicação
