+++ projetoa3/src/main/java/br/com/projetoa3/dto/PacientesRequestDTO.java
@@ -1,0 +1,4 @@
+package br.com.projetoa3.dto;
+
+public record PacientesRequestDTO(String nome, Integer idade, String pressao, String temperatura) {
+}
