+++ projetoa3/src/main/java/br/com/projetoa3/dto/PacientesResponseDTO.java
@@ -1,0 +1,10 @@
+package br.com.projetoa3.dto;
+
+import br.com.projetoa3.entity.Pacientes;
+
+public record PacientesResponseDTO(Long id, String nome, Integer idade, String pressao, String temperatura) {
+    public  PacientesResponseDTO(Pacientes pacientes){
+        this(pacientes.getId(), pacientes.getNome(), pacientes.getIdade(), pacientes.getPressao(), pacientes.getTemperatura());
+
+    }
+}
