+++ projetoa3/src/main/java/br/com/projetoa3/entity/Pacientes.java
@@ -1,25 +1,32 @@
-package br.com.projetoa3.projetoa3.pacientesEntity;
+package br.com.projetoa3.entity;
 
+import br.com.projetoa3.dto.PacientesRequestDTO;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
-@Table (name = "pacientes")
-@Entity (name = "pacientes")
 @Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@Entity
+@Table(name = "pacientes")
 @EqualsAndHashCode(of = "id")
 public class Pacientes {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Long id;
+
+    @Column(name = "nome_do_paciente")
     private String nome;
+
+    @Column(name = "idade_do_paciente")
     private Integer idade;
-    private Integer pressao;
-    private Integer temperatura;
+
+    @Column(name = "pressao_do_paciente")
+    private String pressao;
+
+    @Column(name = "temperatura_do_paciente")
+    private String temperatura;
 
     public  Pacientes(PacientesRequestDTO data){
         this.nome = data.nome();
