--- projetoa3/src/main/java/br/com/projetoa3/projetoa3/pacientesEntity/PacientesRequestDTO.java
@@ -1,4 +1,0 @@
-package br.com.projetoa3.projetoa3.pacientesEntity;
-
-public record PacientesRequestDTO(String nome, Integer idade, Integer pressao, Integer temperatura) {
-}
