--- projetoa3/src/main/java/br/com/projetoa3/projetoa3/pacientesEntity/PacientesResponseDTO.java
@@ -1,8 +1,0 @@
-package br.com.projetoa3.projetoa3.pacientesEntity;
-
-public record PacientesResponseDTO(Long id, String nome, Integer idade, Integer pressao, Integer temperatura) {
-    public  PacientesResponseDTO(Pacientes pacientes){
-        this(pacientes.getId(), pacientes.getNome(), pacientes.getTemperatura(), pacientes.getPressao(), pacientes.getIdade());
-
-    }
-}
