+++ projetoa3/src/main/java/br/com/projetoa3/repository/ManchesterRepository.java
@@ -1,5 +1,6 @@
-package br.com.projetoa3.projetoa3.pacientesEntity;
+package br.com.projetoa3.repository;
 
+import br.com.projetoa3.entity.Pacientes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ManchesterRepository extends JpaRepository<Pacientes, Long>{
