+++ projetoa3/src/main/resources/application.properties
@@ -1,3 +1,5 @@
-spring.datasource.url=jdbc:postgresql://localhost:5432/
-spring.datasource.username=root
+spring.datasource.url=jdbc:postgresql://localhost:5432/projeto_manchester
+spring.datasource.username=postgres
 spring.datasource.password=root
+
+spring.jpa.hibernate.ddl-auto=update
